refactor: organize push domain entities and types
