--- src/main/java/devkor/com/teamcback/domain/notification/entity/NotificationType.java
@@ -1,5 +1,0 @@
-package devkor.com.teamcback.domain.notification.entity;
-
-public enum NotificationType {
-    GENERAL
-}
