+++ src/main/java/devkor/com/teamcback/domain/notification/entity/PushMessage.java
@@ -1,5 +1,6 @@
 package devkor.com.teamcback.domain.notification.entity;
 
+import devkor.com.teamcback.domain.notification.entity.type.PushMessageStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
