--- src/main/java/devkor/com/teamcback/domain/notification/entity/PushMode.java
@@ -1,6 +1,0 @@
-package devkor.com.teamcback.domain.notification.entity;
-
-public enum PushMode {
-    TEST,
-    ACTUAL
-}
