+++ src/main/java/devkor/com/teamcback/domain/notification/entity/type/NotificationType.java
@@ -1,0 +1,5 @@
+package devkor.com.teamcback.domain.notification.entity.type;
+
+public enum NotificationType {
+    GENERAL
+}
