+++ src/main/java/devkor/com/teamcback/domain/notification/entity/type/PushMessageStatus.java
@@ -1,4 +1,4 @@
-package devkor.com.teamcback.domain.notification.entity;
+package devkor.com.teamcback.domain.notification.entity.type;
 
 public enum PushMessageStatus {
     QUEUED,
