+++ src/main/java/devkor/com/teamcback/domain/notification/entity/type/PushMode.java
@@ -1,0 +1,6 @@
+package devkor.com.teamcback.domain.notification.entity.type;
+
+public enum PushMode {
+    TEST,
+    ACTUAL
+}
