+++ src/main/java/devkor/com/teamcback/domain/notification/service/PushPayloadFactory.java
@@ -3,8 +3,8 @@ package devkor.com.teamcback.domain.notification.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import devkor.com.teamcback.domain.notification.dto.payload.PushPayload;
-import devkor.com.teamcback.domain.notification.entity.AppVariant;
-import devkor.com.teamcback.domain.notification.entity.PushActionType;
+import devkor.com.teamcback.domain.notification.entity.type.AppVariant;
+import devkor.com.teamcback.domain.notification.entity.type.PushActionType;
 import devkor.com.teamcback.global.exception.exception.GlobalException;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
