+++ src/main/java/devkor/com/teamcback/domain/notification/service/PushTargetResolver.java
@@ -1,8 +1,8 @@
 package devkor.com.teamcback.domain.notification.service;
 
-import devkor.com.teamcback.domain.notification.entity.AppVariant;
+import devkor.com.teamcback.domain.notification.entity.type.AppVariant;
 import devkor.com.teamcback.domain.notification.entity.PushInstallation;
-import devkor.com.teamcback.domain.notification.entity.PushTargetType;
+import devkor.com.teamcback.domain.notification.entity.type.PushTargetType;
 import devkor.com.teamcback.domain.notification.repository.PushInstallationRepository;
 import devkor.com.teamcback.global.exception.exception.GlobalException;
 import java.util.LinkedHashMap;
